+++ src/main/java/roomescape/global/exception/advice/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.global.exception.advice;
 
 import roomescape.global.exception.base.BusinessException;
