+++ src/main/java/roomescape/global/exception/advice/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.global.exception.advice;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
