+++ src/main/java/roomescape/global/exception/support/BusinessExceptionMappingJackson2HttpMessageConverter.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.global.exception.support;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
